Adjust dying max based on current doomed value
