+++ src/styles/actor/sidebar/_index.scss
@@ -119,9 +119,14 @@ aside {
         a.dots {
             display: flex;
             flex-direction: column;
+            gap: 2px;
 
             > * {
               cursor: pointer;
+            }
+            .pips {
+                display: flex;
+                justify-content: center;
             }
         }
         .data-value {
